Database (#16)
feature: postgres database

+++ src/libserver/Util.cpp
@@ -70,7 +70,7 @@ WinFileTime UnixTimeToFileTime(
   const uint64_t convertedUnixTime = Int32x32To64(unixTime, 10000000) + 116444736000000000;
   return WinFileTime {
     .dwLowDateTime = static_cast<uint32_t>(convertedUnixTime),
-    .dwHighDateTime = convertedUnixTime >> 32
+    .dwHighDateTime = static_cast<uint32_t>(convertedUnixTime >> 32)
   };
 }
